==> Figure_2/Panel_D&F/1_generate_filtered_genes_lists.ipynb <==
import glob
import pandas as pd

sort_DEGs_all_results_list = [f for f in glob.glob("DEGs/SS2/met_vs_primary/*.csv")]

sort_DEGs_all_results_list

final_met_genes = pd.DataFrame()
final_tumor_genes = pd.DataFrame()

for i in sort_DEGs_all_results_list:
    temp_met_df = pd.DataFrame()
    temp_tumor_df = pd.DataFrame()
    tumor_ID = i.split('.')[0].split('MAST_')[1]
    df = pd.read_csv(i,index_col = 0)
    df = df[df['p_val'] < 0.05]
    met_genes = df[df['avg_log2FC'] > 0.5].sort_values(by='avg_log2FC', ascending=False).index.tolist()
    tumor_genes = df[df['avg_log2FC'] < -0.5].sort_values(by='avg_log2FC', ascending=False).index.tolist()
    
    temp_met_df[tumor_ID] = met_genes
    temp_tumor_df[tumor_ID] = tumor_genes
    
    
    final_met_genes = pd.concat([final_met_genes,temp_met_df], ignore_index=False, axis=1)
    final_tumor_genes = pd.concat([final_tumor_genes,temp_tumor_df], ignore_index=False, axis=1)
    

final_met_genes

final_tumor_genes

final_met_genes.to_csv('SS2_met_vs_primary_met_genes_list.csv')
final_tumor_genes.to_csv('SS2_met_vs_primary_tumor_genes_list.csv')
