==> Figure_2/Panel_D&F/3_making_files_for_complexheatmap_R_package.ipynb <==
# importing python modules
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import scanpy as sc
from anndata import read_h5ad
from anndata import read_csv
import anndata
from plotnine import * 
from plotnine.data import mtcars
import venn
from scipy import stats
import seaborn as sns
import math
from scipy.stats import pearsonr
from sklearn.metrics import roc_curve, auc, roc_auc_score
import abc
import batchglm.api as glm
import logging
import patsy
import glob
from random import sample
import scipy.sparse
from typing import Union, Dict, Tuple, List, Set
import math
from anndata import AnnData
from typing import Optional, Union, Mapping  # Special
from typing import Sequence, Collection, Iterable  # ABCs
_VarNames = Union[str, Sequence[str]]

%matplotlib inline
sc.logging.print_header()

# Anndata to DF function

from pandas.api.types import is_categorical_dtype

def sanitize_anndata(adata):
    adata._sanitize()

def sc_prepare_dataframe(
    adata: AnnData,
    var_names: Union[_VarNames, Mapping[str, _VarNames]],
    groupby: Optional[str] = None,
    use_raw: Optional[bool] = None,
    log: bool = False,
    num_categories: int = 7,
    layer=None,
    gene_symbols: Optional[str] = None,
):
    """
    Given the anndata object, prepares a data frame in which the row index are the categories
    defined by group by and the columns correspond to var_names.
    Parameters
    ----------
    adata
        Annotated data matrix.
    var_names
        `var_names` should be a valid subset of  `adata.var_names`.
    groupby
        The key of the observation grouping to consider. It is expected that
        groupby is a categorical. If groupby is not a categorical observation,
        it would be subdivided into `num_categories`.
    use_raw
        Use `raw` attribute of `adata` if present.
    log
        Use the log of the values
    num_categories
        Only used if groupby observation is not categorical. This value
        determines the number of groups into which the groupby observation
        should be subdivided.
    gene_symbols
        Key for field in .var that stores gene symbols.
    Returns
    -------
    Tuple of `pandas.DataFrame` and list of categories.
    """
    from scipy.sparse import issparse

    sanitize_anndata(adata)
    if use_raw is None and adata.raw is not None:
        use_raw = True
    if isinstance(var_names, str):
        var_names = [var_names]

    if groupby is not None:
        if groupby not in adata.obs_keys():
            raise ValueError(
                'groupby has to be a valid observation. '
                f'Given {groupby}, valid observations: {adata.obs_keys()}'
            )

    if gene_symbols is not None and gene_symbols in adata.var.columns:
        # translate gene_symbols to var_names
        # slow method but gives a meaningful error if no gene symbol is found:
        translated_var_names = []
        for symbol in var_names:
            if symbol not in adata.var[gene_symbols].values:
                logg.error(
                    f"Gene symbol {symbol!r} not found in given "
                    f"gene_symbols column: {gene_symbols!r}"
                )
                return
            translated_var_names.append(
                adata.var[adata.var[gene_symbols] == symbol].index[0]
            )
        symbols = var_names
        var_names = translated_var_names
    if layer is not None:
        if layer not in adata.layers.keys():
            raise KeyError(
                f'Selected layer: {layer} is not in the layers list. '
                f'The list of valid layers is: {adata.layers.keys()}'
            )
        matrix = adata[:, var_names].layers[layer]
    elif use_raw:
        matrix = adata.raw[:, var_names].X
    else:
        matrix = adata[:, var_names].X

    if issparse(matrix):
        matrix = matrix.toarray()
    if log:
        matrix = np.log1p(matrix)

    obs_tidy = pd.DataFrame(matrix, columns=var_names)
    if groupby is None:
        groupby = ''
        categorical = pd.Series(np.repeat('', len(obs_tidy))).astype('category')
    else:
        if not is_categorical_dtype(adata.obs[groupby]):
            # if the groupby column is not categorical, turn it into one
            # by subdividing into  `num_categories` categories
            categorical = pd.cut(adata.obs[groupby], num_categories)
        else:
            categorical = adata.obs[groupby]

    obs_tidy.set_index(categorical, groupby, inplace=True)
    if gene_symbols is not None:
        # translate the column names to the symbol names
        obs_tidy.rename(
            columns=dict([(var_names[x], symbols[x]) for x in range(len(var_names))]),
            inplace=True,
        )
    categories = obs_tidy.index.categories

    return categories, obs_tidy


# making files for panel H heatmap

adata_processed = sc.read_h5ad('SS2_processed.h5ad')

adata_processed_tumor = adata_processed[adata_processed.obs['sort'] == 'Tumor'].copy()
adata_processed_met = adata_processed[adata_processed.obs['sort'] == 'Metastatic'].copy()

final_met_genes = pd.read_csv('SS2_met_vs_primary_met_genes_list.csv')
final_tumor_genes = pd.read_csv('SS2_met_vs_primary_tumor_genes_list.csv')

low_met_tumors = ['HCI005', 'H3204','H4272']
intermediate_met_tumors = ['HCI009', 'HCI011', 'HCI001']
high_met_tumors = ['H5097', 'J2036', 'J53353', 'HCI010']

def get_met_and_tumor_genes(tumors_list):
    met_genes_list = []
    tumor_genes_list = []
    for i in tumors_list:
        met_genes = [x for x in final_met_genes[i].tolist() if str(x) != 'nan']
        tumor_genes =[x for x in final_tumor_genes[i].tolist() if str(x) != 'nan']
        
        met_genes_list = met_genes_list + met_genes
        tumor_genes_list = tumor_genes_list + tumor_genes
        
    return met_genes_list, tumor_genes_list

low_group_met_genes, low_group_tumor_genes = get_met_and_tumor_genes(low_met_tumors)
intermediate_group_met_genes, intermediate_group_tumor_genes = get_met_and_tumor_genes(intermediate_met_tumors)
high_group_met_genes, high_group_tumor_genes = get_met_and_tumor_genes(high_met_tumors)

low_group_met_genes_overlap = list(set([x for x in low_group_met_genes if low_group_met_genes.count(x) >= 2]))
len(low_group_met_genes_overlap)

intermediate_group_met_genes_overlap = list(set([x for x in intermediate_group_met_genes if intermediate_group_met_genes.count(x) >= 2]))
len(intermediate_group_met_genes_overlap)

high_group_met_genes_overlap = list(set([x for x in high_group_met_genes if high_group_met_genes.count(x) >= 2]))
len(high_group_met_genes_overlap)

low_group_tumor_genes_overlap = list(set([x for x in low_group_tumor_genes if low_group_tumor_genes.count(x) >= 2]))
len(low_group_tumor_genes_overlap)

intermediate_group_tumor_genes_overlap = list(set([x for x in intermediate_group_tumor_genes if intermediate_group_tumor_genes.count(x) >= 2]))
len(intermediate_group_tumor_genes_overlap)

high_group_tumor_genes_overlap = list(set([x for x in high_group_tumor_genes if high_group_tumor_genes.count(x) >= 2]))
high_group_tumor_genes_overlap.sort()
len(high_group_tumor_genes_overlap)


all_genes = low_group_tumor_genes_overlap + intermediate_group_tumor_genes_overlap + high_group_tumor_genes_overlap + low_group_met_genes_overlap + intermediate_group_met_genes_overlap + high_group_met_genes_overlap



len(all_genes)

tumor_genes_temp = low_group_tumor_genes_overlap + intermediate_group_tumor_genes_overlap + high_group_tumor_genes_overlap

print(len(tumor_genes_temp), len(set(tumor_genes_temp)))

tumor_genes = []
for i in tumor_genes_temp:
    if i not in tumor_genes:
        tumor_genes.append(i)

print(len(tumor_genes), len(set(tumor_genes)))

met_genes = []
for i in low_group_met_genes_overlap + intermediate_group_met_genes_overlap + high_group_met_genes_overlap:
    if i not in tumor_genes and i not in met_genes:
        met_genes.append(i)

print(len(met_genes), len(set(met_genes)))

tumor_mean_expression_df_tumor_genes = sc_prepare_dataframe(adata_processed_tumor, tumor_genes, 
                                                groupby='Tumor_ID',use_raw=False)

tumor_mean_expression_df_met_genes = sc_prepare_dataframe(adata_processed_tumor, met_genes, 
                                                groupby='Tumor_ID',use_raw=False)

tumor_mean_expression_df_tumor_genes = tumor_mean_expression_df_tumor_genes[1].groupby(level=0).mean()
tumor_mean_expression_df_tumor_genes.index = [i+'_Tumor' for i in tumor_mean_expression_df_tumor_genes.index]
tumor_mean_expression_df_tumor_genes = tumor_mean_expression_df_tumor_genes.T
tumor_mean_expression_df_tumor_genes

tumor_mean_expression_df_met_genes = tumor_mean_expression_df_met_genes[1].groupby(level=0).mean()
tumor_mean_expression_df_met_genes.index = [i+'_Tumor' for i in tumor_mean_expression_df_met_genes.index]
tumor_mean_expression_df_met_genes = tumor_mean_expression_df_met_genes.T
tumor_mean_expression_df_met_genes

met_mean_expression_df_tumor_genes = sc_prepare_dataframe(adata_processed_met, tumor_genes, 
                                              groupby='Tumor_ID',use_raw=False)

met_mean_expression_df_met_genes = sc_prepare_dataframe(adata_processed_met, met_genes, 
                                              groupby='Tumor_ID',use_raw=False)

met_mean_expression_df_tumor_genes = met_mean_expression_df_tumor_genes[1].groupby(level=0).mean()
met_mean_expression_df_tumor_genes.index = [i+'_Metastatic' for i in met_mean_expression_df_tumor_genes.index]
met_mean_expression_df_tumor_genes = met_mean_expression_df_tumor_genes.T
met_mean_expression_df_tumor_genes

met_mean_expression_df_met_genes = met_mean_expression_df_met_genes[1].groupby(level=0).mean()
met_mean_expression_df_met_genes.index = [i+'_Metastatic' for i in met_mean_expression_df_met_genes.index]
met_mean_expression_df_met_genes = met_mean_expression_df_met_genes.T
met_mean_expression_df_met_genes

global_mean_expression_df_tumor_genes = sc_prepare_dataframe(adata_processed, tumor_genes
                                                 , groupby='sort',use_raw=False)

global_mean_expression_df_met_genes = sc_prepare_dataframe(adata_processed, met_genes
                                                 , groupby='sort',use_raw=False)


global_mean_expression_df_tumor_genes = global_mean_expression_df_tumor_genes[1].groupby(level=0).mean()
global_mean_expression_df_tumor_genes.index = ['global_'+i for i in global_mean_expression_df_tumor_genes.index]
global_mean_expression_df_tumor_genes = global_mean_expression_df_tumor_genes.T
global_mean_expression_df_tumor_genes

global_mean_expression_df_met_genes = global_mean_expression_df_met_genes[1].groupby(level=0).mean()
global_mean_expression_df_met_genes.index = ['global_'+i for i in global_mean_expression_df_met_genes.index]
global_mean_expression_df_met_genes = global_mean_expression_df_met_genes.T
global_mean_expression_df_met_genes

combined_tumor_tumor_genes = pd.concat([tumor_mean_expression_df_tumor_genes, global_mean_expression_df_tumor_genes[['global_Tumor']]],axis=1)
combined_tumor_tumor_genes

combined_tumor_met_genes = pd.concat([tumor_mean_expression_df_met_genes, global_mean_expression_df_met_genes[['global_Tumor']]],axis=1)
combined_tumor_met_genes

combined_met_tumor_genes = pd.concat([met_mean_expression_df_tumor_genes, global_mean_expression_df_tumor_genes[['global_Metastatic']]],axis=1)
combined_met_tumor_genes

combined_met_met_genes = pd.concat([met_mean_expression_df_met_genes, global_mean_expression_df_met_genes[['global_Metastatic']]],axis=1)
combined_met_met_genes

new_order_tumor = ['HCI005_Tumor','H3204_Tumor','H4272_Tumor','HCI009_Tumor',
            'HCI011_Tumor','HCI001_Tumor','H5097_Tumor','J2036_Tumor','J53353_Tumor','HCI010_Tumor','global_Tumor']

new_order_met = ['HCI005_Metastatic','H3204_Metastatic','H4272_Metastatic','HCI009_Metastatic',
            'HCI011_Metastatic','HCI001_Metastatic','H5097_Metastatic','J2036_Metastatic','J53353_Metastatic','HCI010_Metastatic','global_Metastatic']

combined_tumor_tumor_genes = combined_tumor_tumor_genes[new_order_tumor]
combined_tumor_met_genes = combined_tumor_met_genes[new_order_tumor]

combined_met_tumor_genes = combined_met_tumor_genes[new_order_met]
combined_met_met_genes = combined_met_met_genes[new_order_met]

combined_tumor_tumor_genes.to_csv('tumor_mean_tumor_genes_gc.csv')
combined_tumor_met_genes.to_csv('tumor_mean_met_genes_gc.csv')

combined_met_tumor_genes.to_csv('met_mean_tumor_genes_gc.csv')
combined_met_met_genes.to_csv('met_mean_met_genes_gc.csv')

metadata = pd.DataFrame(index=new_order_tumor)
metadata['Tumor_ID'] = [i.split('_')[0] for i in metadata.index]
metadata['sort'] = [i.split('_')[1] for i in metadata.index]
for i in metadata.index:
    tumor_id = metadata.loc[i, 'Tumor_ID']
    if tumor_id in low_met_tumors:
        metadata.loc[i, 'metastatic_potential_group'] = 'low'
    elif tumor_id in intermediate_met_tumors+['H5471']:
        metadata.loc[i, 'metastatic_potential_group'] = 'moderate'
    elif tumor_id in high_met_tumors:
        metadata.loc[i, 'metastatic_potential_group'] = 'high'
    else:
        metadata.loc[i, 'metastatic_potential_group'] = 'all'
metadata

metadata.to_csv('tumor_metadata.csv')

metadata = pd.DataFrame(index=new_order_met)
metadata['Tumor_ID'] = [i.split('_')[0] for i in metadata.index]
metadata['sort'] = [i.split('_')[1] for i in metadata.index]
for i in metadata.index:
    tumor_id = metadata.loc[i, 'Tumor_ID']
    if tumor_id in low_met_tumors:
        metadata.loc[i, 'metastatic_potential_group'] = 'low'
    elif tumor_id in intermediate_met_tumors+['H5471']:
        metadata.loc[i, 'metastatic_potential_group'] = 'moderate'
    elif tumor_id in high_met_tumors:
        metadata.loc[i, 'metastatic_potential_group'] = 'high'
    else:
        metadata.loc[i, 'metastatic_potential_group'] = 'all'
metadata

metadata.to_csv('met_metadata.csv')

gene_metadata = pd.DataFrame()
gene_metadata['gene'] = tumor_genes
gene_metadata = gene_metadata.set_index('gene')
gene_metadata['sort'] = 'tumor'
for i in tumor_genes:
    if i in low_group_tumor_genes_overlap:
        gene_metadata.loc[i, 'group'] = 'low'
    elif i in intermediate_group_tumor_genes_overlap:
        gene_metadata.loc[i, 'group'] = 'moderate'
    elif i in high_group_tumor_genes_overlap:
        gene_metadata.loc[i, 'group'] = 'high'

gene_metadata

gene_metadata.to_csv('tumor_genes_metadata.csv')

gene_metadata = pd.DataFrame()
gene_metadata['gene'] = met_genes
gene_metadata = gene_metadata.set_index('gene')
gene_metadata['sort'] = 'tumor'
for i in met_genes:
    if i in low_group_met_genes_overlap:
        gene_metadata.loc[i, 'group'] = 'low'
    elif i in intermediate_group_met_genes_overlap:
        gene_metadata.loc[i, 'group'] = 'moderate'
    elif i in high_group_met_genes_overlap:
        gene_metadata.loc[i, 'group'] = 'high'

gene_metadata

gene_metadata.to_csv('met_genes_metadata.csv')
